refactor(web): migrate type-converter to feature folder
Also relax the assembleMessages invariant test from toEqual(central) to
toMatchObject(central): once a tool is migrated, fragments only add to
the trimmed central catalog, so central is a subset of the assembled
result at every migration step.
