+++ apps/web/src/messages/zh-TW.json
@@ -44,7 +44,6 @@
   },
   "Tools": {
     "object-flatten": { "title": "物件壓平 / 還原", "description": "把巢狀物件壓平成點路徑鍵,並可還原。" },
-    "type-converter": { "title": "資料型別轉換器", "description": "在字串、數字、布林、日期之間轉換值。" },
     "data-filter-tester": { "title": "JSONPath 篩選測試器", "description": "對範例資料即時執行 @rfjs/data-filter 條件。" },
     "jwt-decoder": { "title": "JWT 解碼器", "description": "解碼 JWT 的 header 與 payload，並顯示即時有效期狀態。" },
     "jsonb-query-generator": { "title": "篩選 → JSONB SQL", "description": "從篩選 metadata 產生 PostgreSQL JSONB 查詢。" },
@@ -56,8 +55,6 @@
   "ToolUI": {
     "input": "輸入",
     "output": "輸出",
-    "inputValue": "值",
-    "targetType": "目標型別",
     "jsonInput": "JSON",
     "copy": "複製",
     "data": "資料",
