+++ apps/web/src/tools/index.ts
@@ -2,8 +2,10 @@ import type { ComponentType } from "react";
 
 import type { ToolModule } from "./types";
 
+import { tool as typeConverter } from "./type-converter";
+
 // As each tool is migrated, add its descriptor to this array.
-export const toolModules: ToolModule[] = [];
+export const toolModules: ToolModule[] = [typeConverter];
 
 export const TOOL_COMPONENTS: Record<string, ComponentType> = Object.fromEntries(
   toolModules.map((t) => [t.id, t.Component]),
