+++ apps/web/src/tools/type-converter/index.ts
@@ -1,0 +1,5 @@
+import type { ToolModule } from "@/tools/types";
+
+import { TypeConverter } from "./ui";
+
+export const tool: ToolModule = { id: "type-converter", Component: TypeConverter };
